requirements-dev.txt
-r requirements.txt
pytest
coverage
requests-mock